Renaming reducer from objectstore to attachments
