+++ frontend/src/app/store/reducers.ts
@@ -8,7 +8,7 @@ import complaints from "./reducers/complaints";
 import offices from "./reducers/office";
 import codeTables from "./reducers/code-table";
 import complaintLocations from "./reducers/complaint-locations";
-import attachments from "./reducers/objectstore";
+import attachments from "./reducers/attachments";
 
 const appPersistConfig = {
 	key: "app",
